skills/meeting_assistant/requirements.txt
google-generativeai
requests
watchdog
